+++ Game/Ship_Interactions.cpp
@@ -15,6 +15,7 @@
 
 #include <algorithm>
 #include <cassert>
+#include <cmath>
 #include <limits>
 
 namespace Physics {
@@ -1337,7 +1338,7 @@ std::optional<vec2f> Ship::FindSuitableLightningTarget() const
                         {
                             // Height of to-be-inserted point is augmented based on distance from the point
                             float const distance = (candidatePos - pos).length();
-                            float const actualPosY = pos.y + std::max(floorf(distance / 3.0f), 5.0f);
+                            float const actualPosY = pos.y + std::max(floor(distance / 3.0f), 5.0f);
                             return candidatePos.y > actualPosY;
                         }),
                     pos);
